# Use a stricter form of search for extremely short queries