@sgrif 2019-05-22
I've finally tracked down where the weird short search queries are coming from -- It's from IDE autocompletion. I think these results are fine for that, or we can switch to prefix matching which will still hit our index. We should probably look into adding a more optimized endpoint for this use case