@jtgeibel 2019-05-22
> It's from IDE autocompletion.

That's interesting.  I hadn't noticed before that the RLS did autocompletion of crate names in `Cargo.toml`.  I should be possible for the RLS to do this locally from the index, but these changes look good to me.

@bors r+