@sgrif 2019-05-23
Yeah I think rls uses the index, this is from intellij which probably
should use the index as well. I opened a pr so they send a UA at least

On Wed, May 22, 2019, 6:08 PM bors <notifications@github.com> wrote:

> Merged #1752 <https://github.com/rust-lang/crates.io/pull/1752> into
> master.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/rust-lang/crates.io/pull/1752?email_source=notifications&email_token=AALVMKZ26AEKS7JVRVUCMHLPWXN7DA5CNFSM4HOXSUI2YY3PNVWWK3TUL52HS4DFWZEXG43VMVCXMZLOORHG65DJMZUWGYLUNFXW5KTDN5WW2ZLOORPWSZGORS4LR3A#event-2360916204>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AALVMK2MLQMOHZMUNM3442LPWXN7DANCNFSM4HOXSUIQ>
> .
>